requirements.txt
kornia==0.7.2
matplotlib==3.3.0
numpy==1.24.3
onnx==1.16.1
onnxruntime==1.18.0
opencv_python==4.9.0.80
Pillow==10.3.0
tensorflow==2.13.1
torch==1.12.1+cu113
